virtual-diary-app-main-2/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
gspread
oauth2client
streamlit
python-dotenv
